requirements.txt
google-generativeai
chromadb
pdfplumber
python-dotenv
pandas